+++ src/test/java/roomescape/service/reservation/ReservationCreateServiceTest.java
@@ -1,65 +1,21 @@
 package roomescape.service.reservation;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.jdbc.Sql;
-import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberRepository;
-import roomescape.domain.member.Role;
 import roomescape.domain.reservation.ReservationStatus;
-import roomescape.domain.reservationdetail.ReservationDetail;
-import roomescape.domain.reservationdetail.ReservationDetailRepository;
-import roomescape.domain.schedule.ReservationDate;
-import roomescape.domain.schedule.ReservationTime;
-import roomescape.domain.schedule.ReservationTimeRepository;
-import roomescape.domain.schedule.Schedule;
-import roomescape.domain.theme.Theme;
-import roomescape.domain.theme.ThemeRepository;
 import roomescape.exception.InvalidReservationException;
 import roomescape.service.reservation.dto.AdminReservationRequest;
 import roomescape.service.reservation.dto.ReservationRequest;
 import roomescape.service.reservation.dto.ReservationResponse;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
-@Sql("/truncate-with-time-and-theme.sql")
-class ReservationCreateServiceTest {
+class ReservationCreateServiceTest extends ReservationServiceTest {
     @Autowired
     private ReservationCreateService reservationCreateService;
-    @Autowired
-    private ReservationTimeRepository reservationTimeRepository;
-    @Autowired
-    private ThemeRepository themeRepository;
-    @Autowired
-    private MemberRepository memberRepository;
-    @Autowired
-    private ReservationDetailRepository reservationDetailRepository;
-    private ReservationDetail reservationDetail;
-    private Theme theme;
-    private Member admin;
-    private Member member;
-    private Member anotherMember;
-
-    @BeforeEach
-    void setUp() {
-        ReservationDate reservationDate = ReservationDate.of(LocalDate.now().plusDays(1));
-        ReservationTime reservationTime = reservationTimeRepository.save(new ReservationTime(LocalTime.of(10, 0)));
-        theme = themeRepository.save(new Theme("레벨2 탈출", "우테코 레벨2를 탈출하는 내용입니다.",
-                "https://i.pinimg.com/236x/6e/bc/46/6ebc461a94a49f9ea3b8bbe2204145d4.jpg"));
-        admin = memberRepository.save(new Member("admin", "admin@email.com", "admin123", Role.ADMIN));
-        member = memberRepository.save(new Member("lini", "lini@email.com", "lini123", Role.GUEST));
-        anotherMember = memberRepository.save(new Member("pedro", "pedro@email.com", "pedro123", Role.GUEST));
-        reservationDetail = reservationDetailRepository.save(new ReservationDetail(new Schedule(reservationDate, reservationTime), theme));
-    }
 
     @DisplayName("어드민이 새로운 예약을 저장한다.")
     @Test
