+++ src/test/java/roomescape/service/reservation/ReservationReadServiceTest.java
@@ -1,0 +1,80 @@
+package roomescape.service.reservation;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.jdbc.Sql;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationRepository;
+import roomescape.domain.reservation.ReservationStatus;
+import roomescape.service.reservation.dto.ReservationFilterRequest;
+import roomescape.service.reservation.dto.ReservationResponse;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class ReservationReadServiceTest extends ReservationServiceTest {
+    @Autowired
+    private ReservationReadService reservationReadService;
+    @Autowired
+    private ReservationRepository reservationRepository;
+
+    @DisplayName("모든 예약 내역을 조회한다.")
+    @Test
+    @Sql({"/truncate-with-time-and-theme.sql", "/insert-past-reservation.sql"})
+    void findAllReservations() {
+        //when
+        List<ReservationResponse> reservations = reservationReadService.findAll();
+
+        //then
+        assertThat(reservations).hasSize(3);
+    }
+
+    @DisplayName("사용자 조건으로 예약 내역을 조회한다.")
+    @Test
+    void findByMember() {
+        //given
+        Reservation reservation = new Reservation(member, reservationDetail, ReservationStatus.RESERVED);
+        reservationRepository.save(reservation);
+        ReservationFilterRequest reservationFilterRequest = new ReservationFilterRequest(member.getId(), null, null,
+                null);
+
+        //when
+        List<ReservationResponse> reservations = reservationReadService.findByCondition(reservationFilterRequest);
+
+        //then
+        assertThat(reservations).hasSize(1);
+    }
+
+    @DisplayName("사용자와 테마 조건으로 예약 내역을 조회한다.")
+    @Test
+    void findByMemberAndTheme() {
+        //given
+        Reservation reservation = new Reservation(member, reservationDetail, ReservationStatus.RESERVED);
+        reservationRepository.save(reservation);
+        long notMemberThemeId = theme.getId() + 1;
+        ReservationFilterRequest reservationFilterRequest = new ReservationFilterRequest(member.getId(),
+                notMemberThemeId, null, null);
+
+        //when
+        List<ReservationResponse> reservations = reservationReadService.findByCondition(reservationFilterRequest);
+
+        //then
+        assertThat(reservations).isEmpty();
+    }
+
+    @DisplayName("관리자가 id로 예약을 삭제한다.")
+    @Test
+    void deleteReservationById() {
+        //given
+        Reservation reservation = new Reservation(admin, reservationDetail, ReservationStatus.RESERVED);
+        Reservation target = reservationRepository.save(reservation);
+
+        //when
+        reservationRepository.deleteById(target.getId());
+
+        //then
+        assertThat(reservationReadService.findAll()).isEmpty();
+    }
+}
