+++ src/test/java/roomescape/service/reservation/ReservationServiceTest.java
@@ -1,18 +1,12 @@
 package roomescape.service.reservation;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.domain.dto.ReservationWithRank;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
 import roomescape.domain.member.Role;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationRepository;
-import roomescape.domain.reservation.ReservationStatus;
 import roomescape.domain.reservationdetail.ReservationDetail;
 import roomescape.domain.reservationdetail.ReservationDetailRepository;
 import roomescape.domain.schedule.ReservationDate;
@@ -21,24 +15,13 @@ import roomescape.domain.schedule.ReservationTimeRepository;
 import roomescape.domain.schedule.Schedule;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeRepository;
-import roomescape.exception.InvalidReservationException;
-import roomescape.service.reservation.dto.ReservationFilterRequest;
-import roomescape.service.reservation.dto.ReservationResponse;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @Sql("/truncate-with-time-and-theme.sql")
-class ReservationServiceTest {
-    @Autowired
-    private ReservationService reservationService;
-    @Autowired
-    private ReservationRepository reservationRepository;
+public abstract class ReservationServiceTest {
     @Autowired
     private ReservationTimeRepository reservationTimeRepository;
     @Autowired
@@ -47,11 +30,11 @@ class ReservationServiceTest {
     private MemberRepository memberRepository;
     @Autowired
     private ReservationDetailRepository reservationDetailRepository;
-    private ReservationDetail reservationDetail;
-    private Theme theme;
-    private Member admin;
-    private Member member;
-    private Member anotherMember;
+    protected ReservationDetail reservationDetail;
+    protected Theme theme;
+    protected Member admin;
+    protected Member member;
+    protected Member anotherMember;
 
     @BeforeEach
     void setUp() {
@@ -64,94 +47,5 @@ class ReservationServiceTest {
         anotherMember = memberRepository.save(new Member("pedro", "pedro@email.com", "pedro123", Role.GUEST));
         reservationDetail = reservationDetailRepository.save(new ReservationDetail(new Schedule(reservationDate, reservationTime), theme));
     }
-
-    @DisplayName("모든 예약 내역을 조회한다.")
-    @Test
-    @Sql({"/truncate-with-time-and-theme.sql", "/insert-past-reservation.sql"})
-    void findAllReservations() {
-        //when
-        List<ReservationResponse> reservations = reservationService.findAll();
-
-        //then
-        assertThat(reservations).hasSize(3);
-    }
-
-    @DisplayName("사용자 조건으로 예약 내역을 조회한다.")
-    @Test
-    void findByMember() {
-        //given
-        Reservation reservation = new Reservation(member, reservationDetail, ReservationStatus.RESERVED);
-        reservationRepository.save(reservation);
-        ReservationFilterRequest reservationFilterRequest = new ReservationFilterRequest(member.getId(), null, null,
-                null);
-
-        //when
-        List<ReservationResponse> reservations = reservationService.findByCondition(reservationFilterRequest);
-
-        //then
-        assertThat(reservations).hasSize(1);
-    }
-
-    @DisplayName("사용자와 테마 조건으로 예약 내역을 조회한다.")
-    @Test
-    void findByMemberAndTheme() {
-        //given
-        Reservation reservation = new Reservation(member, reservationDetail, ReservationStatus.RESERVED);
-        reservationRepository.save(reservation);
-        long notMemberThemeId = theme.getId() + 1;
-        ReservationFilterRequest reservationFilterRequest = new ReservationFilterRequest(member.getId(),
-                notMemberThemeId, null, null);
-
-        //when
-        List<ReservationResponse> reservations = reservationService.findByCondition(reservationFilterRequest);
-
-        //then
-        assertThat(reservations).isEmpty();
-    }
-
-    @DisplayName("관리자가 id로 예약을 삭제한다.")
-    @Test
-    void deleteReservationById() {
-        //given
-        Reservation reservation = new Reservation(admin, reservationDetail, ReservationStatus.RESERVED);
-        Reservation target = reservationRepository.save(reservation);
-
-        //when
-        reservationService.deleteById(target.getId());
-
-        //then
-        assertThat(reservationService.findAll()).isEmpty();
-    }
-
-    @DisplayName("관리자가 예약을 삭제하고, 예약 대기가 있다면 가장 우선순위가 높은 예약 대기를 예약으로 전환한다.")
-    @Test
-    void deleteThenUpdateReservation() {
-        //given
-        Reservation reservation = new Reservation(admin, reservationDetail, ReservationStatus.RESERVED);
-        Reservation reservation2 = new Reservation(member, reservationDetail, ReservationStatus.WAITING);
-        Reservation reservation3 = new Reservation(anotherMember, reservationDetail, ReservationStatus.WAITING);
-        reservationRepository.save(reservation);
-        reservationRepository.save(reservation2);
-        reservationRepository.save(reservation3);
-
-        //when
-        reservationService.deleteById(reservation.getId());
-
-        //then
-        List<ReservationWithRank> reservations = reservationRepository.findWithRankingByMemberId(member.getId());
-        assertThat(reservations.get(0).getReservation().isReserved()).isTrue();
-    }
-
-    @DisplayName("관리자가 과거 예약을 삭제하려고 하면 예외가 발생한다.")
-    @Test
-    @Sql({"/truncate.sql", "/insert-past-reservation.sql"})
-    void cannotDeleteReservationByIdIfPast() {
-        //given
-        long id = 1;
-
-        //when & then
-        assertThatThrownBy(() -> reservationService.deleteById(id))
-                .isInstanceOf(InvalidReservationException.class)
-                .hasMessage("이미 지난 예약은 삭제할 수 없습니다.");
-    }
 }
+
